--- a/Project.toml
+++ b/Project.toml
@@ -22,9 +22,11 @@ UnitfulAtomic = "a7773ee8-282e-5fa2-be4e-bd808c38a91a"
 
 [weakdeps]
 Molly = "aa0f7f06-fcc0-5ec4-a7f3-a573f33f9c4c"
+ACEfit = "ad31a8ef-59f5-4a01-b543-a85c2f73e95c"
 
 [extensions]
 ACE_Molly_ext = "Molly"
+ACEmd_ACEfit_ext = "ACEfit"
 
 [compat]
 ACE1 = "0.11.5"