--- a/Project.toml
+++ b/Project.toml
@@ -32,6 +32,7 @@ ACEmd_ACEfit_ext = "ACEfit"
 [compat]
 ACE1 = "0.11.5"
 ACE1x = "0.1.8"
+ACEfit = "0.1.4"
 ACEbase = "0.4.3"
 AtomsBase = "0.3"
 CellListMap = "0.8"