--- a/docs/Project.toml
+++ b/docs/Project.toml
@@ -1,4 +1,6 @@
 [deps]
+ACE1x = "5cc4c08c-8782-4a30-af6d-550b302e9707"
+ACEfit = "ad31a8ef-59f5-4a01-b543-a85c2f73e95c"
 ACEmd = "69e0c927-b120-467d-b2b3-5b6842148cf4"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ExtXYZ = "352459e4-ddd7-4360-8937-99dcb397b478"