--- a/docs/make.jl
+++ b/docs/make.jl
@@ -18,6 +18,7 @@ makedocs(;
     pages=[
         "Home" => "index.md",
         "Using ACE in Molly" => "molly.md",
+        "Fitting Potential" => "fitting.jl",
         "Benchmarks" => "benchmark.md"
     ],
 )