--- a/src/api.jl
+++ b/src/api.jl
@@ -460,11 +460,132 @@ end
 
 for ace_method in [ :ace_energy, :ace_forces, :ace_virial]
     @eval begin
-        function $ace_method(basis::ACE1.IPSuperBasis, data; executor=ThreadedEx(), kwargs...)
-            em = Folds.map( basis.BB, executor ) do b
-                $ace_method(b, data; executor=executor, kwargs...)
+        function $ace_method(basis::ACE1.IPSuperBasis, data; kwargs...)
+            em = map( basis.BB ) do b
+                Threads.@spawn $ace_method(b, data; kwargs...)
             end
-            return reduce(vcat, em)
+            final_data = [ fetch(x) for x in em ]
+            return reduce(vcat, final_data)
         end
     end
+end
+
+
+function ace_forces_virial(basis::ACE1.IPSuperBasis, data; kwargs...)
+    em = map( basis.BB ) do b
+        Threads.@spawn ace_forces_virial(b, data; kwargs...)
+    end
+    final_data = [ fetch(x) for x in em ]
+    Farrays = [ x["force"] for x in final_data ]
+    Varrays = [ x["virial"] for x in final_data ]
+    F = reduce(vcat, Farrays)
+    V = reduce(vcat, Varrays)
+    return Dict("force"=>F, "virial"=>V)
+end
+
+# special combination calls for extra speed
+
+function ace_forces_virial(
+    basis::ACE1.IPBasis,
+    at;
+    domain=1:length(at),
+    executor=ThreadedEx(),
+    ntasks=Threads.nthreads(),
+    cutoff_unit=default_length,
+    kwargs...
+)   
+    nlist = neighborlist(at, get_cutoff(basis; cutoff_unit=cutoff_unit) )
+    F_and_V = Folds.sum( collect(chunks(domain, ntasks)), executor ) do (d, _)
+        ace_forces_virial(basis, at, nlist; domain=d)
+    end
+    F = [ Vector(f) for f in eachrow(F_and_V[1])]
+    return Dict("force"=>F, "virial"=>F_and_V[2])
+end
+
+
+function ace_forces_virial(
+    basis::ACE1.IPBasis,
+    at,
+    nlist;
+    domain=1:length(at)
+)
+    f = zeros(SVector{3, Float64}, length(basis), length(at))
+    vir = zeros(SMatrix{3,3,Float64}, length(basis))
+    for i in domain
+        j, R, Z = neigsz(nlist, at, i)
+        dB = ace_evaluate_d(basis, R, Z, _atomic_number(at,i))
+
+        for k in eachindex(j)
+            f[:, j[k]] .-= @view dB[:, k]
+        end
+        for k in eachindex(j)
+            f[:,i] .+= @view dB[:, k]
+        end
+        for ib in 1:length(basis)
+            site_virial = -sum( zip( R, view(dB, ib,:) ) ) do (Rⱼ, dVⱼ)
+                dVⱼ * Rⱼ'
+            end
+            vir[ib] += site_virial
+        end
+    end
+    return [f, vir]
+end
+
+
+function ace_forces_virial(
+    pair_basis::ACE1.PolyPairBasis,
+    at;
+    domain=1:length(at),
+    executor=ThreadedEx(),
+    ntasks=Threads.nthreads(),
+    cutoff_unit=default_length,
+    kwargs...
+)   
+    nlist = neighborlist(at, get_cutoff(pair_basis; cutoff_unit=cutoff_unit) )
+    pair_data = [ (i, j, R)  for (i,j,R) in pairs(nlist) if i in domain ]
+    F_and_V = Folds.sum( collect(chunks(pair_data, ntasks)), executor ) do (d, _)
+        ace_forces_virial(pair_basis, at, pair_data, d)
+    end
+    F = [ Vector(f) for f in eachrow(F_and_V[1])]
+    return Dict("force"=>F, "virial"=>F_and_V[2])
+end
+
+
+function ace_forces_virial(
+    pair_basis::ACE1.PolyPairBasis,
+    at,
+    pair_data,
+    domain
+)
+    f = zeros(SVector{3, Float64}, length(pair_basis), length(at))
+    vir = zeros(SMatrix{3,3,Float64}, length(pair_basis))
+    tmp = ACE1.alloc_temp_d(pair_basis)
+    for I in domain
+        # Pairpotential evaluator from JuLIP.
+        # This is temporary implementation and will be changed in the future
+        i, j, R = pair_data[I]
+        r = norm(R)
+        Zi = _atomic_number(at, i)
+        Zj = _atomic_number(at, j)
+        Ii = ACE1.z2i(pair_basis, Zi)
+        Ij = ACE1.z2i(pair_basis, Zj)
+        dJ = tmp.dJ[Ii, Ij]
+        ACE1.evaluate_d!(tmp.J[Ii, Ij], dJ, tmp.tmpd_J[Ii, Ij], pair_basis.J[Ii, Ij], r, Zi, Zj)
+        idx0 = pair_basis.bidx0[Ii, Ij] 
+        r_tmp = R/(2r)
+        for n = 1:length(pair_basis.J[Ii, Ij])
+            f[idx0 + n, i] +=  dJ[n] * r_tmp
+        end
+        for n = 1:length(pair_basis.J[Ii, Ij])
+            f[idx0 + n, j] -= dJ[n] * r_tmp
+        end
+        # TODO this creates too much allocations
+        # should look more, like R' is an allocation
+        rr_tmp =  R * R'
+        for n = 1:length(pair_basis.J[Ii, Ij])
+            vir[idx0 + n] -= (dJ[n]/(2r)) * rr_tmp
+        end
+        #
+    end
+    return [f, vir]
 end