@cortner This adds training support for `AtomsBase` input. It has only overloads the function in `ACEfit`. Later on we can support for `ACEpotentials` specific functions.

There was some issue with implementing this. I did stumble on a bug that caused `assemble` to give some random results (can be triggered by changing `pmap` in `asseble` to `Folds.map`). This is probably some bug in ACE data allocation, but it can also be in Julia general - I don't understand how the bug works. But I am very concerned of it, as appear somewhere else and possible changes in Julia can cause it to emerge in the existing "working" code.

Here is a example code and comparison to old way

```julia
using ACEmd
using ACEfit
using ACEpotentials

model = acemodel(
    elements = [:Ti, :Al],
	order = 3,
	totaldegree = 6,
	rcut = 5.5,
	Eref = [:Ti => -1586.0195, :Al => -105.5954]
)
basis = model.basis

data_j, _, meta = ACEpotentials.example_dataset("TiAl_tutorial")
weights = Dict( "default" => Dict("E" => 5.0, "F" => 1.0 , "V" => 1.0 ) )
datakeys = (energy_key = "energy", force_key = "force", virial_key = "virial")
train_old = [ACEpotentials.AtomsData(t; weights=weights, v_ref=model.Vref, datakeys...) for t in data_j[1:5:end]]
train_new = [  FlexibleSystem(x.atoms) for x in train_old  ]

# Old style
A, Y, W = ACEfit.assemble(train_old, basis);

# New style
a, y, w = ACEfit.assemble(train_new, basis; energy_default_weight=5, energy_ref=model.Vref)

# Compare results
maximum(abs2, A - a)
maximum(abs2, Y - y)
maximum(abs2, W - w)
```

The new way is way more faster than the old one, between several tens to several hundred (if you don't use parallel processing). The is also a way to make it even faster, but I will skip for now. The main point is that assembly is now faster than solving the linear equation, changing the dynamics of fitting potentials completely.