@cortner 2023-09-16
Hi Teemy - Thank you for exploring this.  I'll will need to find time to go through the code very carefully. In the meantime can you please explain a few things? 
- How does the fitting data enter the new assembly? I don't see this. Does `FlexibleSystem` keep the metadata like reference energies and forces? 
- Similarly, how do the fitting weights enter the new assembly? 
- Can you explain briefly where the speedup comes from?

Regarding the bug, I have some doubts about there being a bug in `ACE1` allocations, it's been in use for many years now, but you never know - as you say changes in Julia or somewhere else might bring it out. But possibly also some incorrect assumptions made about how temporary storage is to be pre-allocated and how outputs are allowed to be used. Either way, I'd be grateful if you can open a separate issue or PR with a reproducable example so we can investigate it. 

CC @wcwitt - would you mind following this discussion as well?