@cortner 2023-09-16
another question - why add the example training data to the repo when you can just get it via `ACEpotentials.example_dataset`?