@tjjarvinen 2023-09-16
> * How does the fitting data enter the new assembly? I don't see this. Does `FlexibleSystem` keep the metadata like reference energies and forces?

Yes, AtomsBase is defined so that you can store whatever data in the structures. The documentation page of this PR has the details. But in short, if `data` is `AtomsBase` structure is should have training data behind keys: `data[:energy]`, `data[:force]` and `data[:virial]`. `ExtXYZ` does have a bug at the moment that makes it fail to read some of this data, like force, when used for AtomsBase structure (works for JuLIP), so I need to submit a PR for it, to get `AtomsBase` interface to work fully.

Fitting weights are also explained in the documentation page. In short, general weights are given as keywords, like in above example. In addition you can give each structure a separate weight parameter that is used only for that structure.

 

> * Can you explain briefly where the speedup comes from?

I am not completely sure. I suspect that it has something to do with the garbage collection issue. Here is a benchmark comparison with the above posted example

```julia
# old impelentation
julia> @benchmark ACEfit.assemble($train_old, $basis)
BenchmarkTools.Trial: 1 sample with 1 evaluation.
 Single result which took 14.657 s (98.42% GC) to evaluate,
 with a memory estimate of 445.05 MiB, over 5253010 allocations.

# new implementation
julia> @benchmark ACEfit.assemble($train_new, $basis; energy_default_weight=$5, energy_ref=$model.Vref) 
BenchmarkTools.Trial: 32 samples with 1 evaluation.
 Range (min … max):  136.020 ms … 180.292 ms  ┊ GC (min … max): 18.63% … 16.11%
 Time  (median):     156.335 ms               ┊ GC (median):    23.22%
 Time  (mean ± σ):   157.097 ms ±   9.635 ms  ┊ GC (mean ± σ):  22.56% ±  2.36%

                     █ ▃▃▃             █   █
  ▇▁▁▁▁▁▁▁▁▇▁▁▇▁▇▁▇▁▁█▇███▁▁▁▇▇▇▁▁▇▇▇▁▇█▁▁▁█▇▇▁▁▁▁▁▁▁▁▁▁▁▁▇▁▁▁▇ ▁
  136 ms           Histogram: frequency by time          180 ms <

 Memory estimate: 1.03 GiB, allocs estimate: 5453582.
```

Use of multiprocessing in the old implementation makes it hard to see, where the difference comes. But it is probably the garbage collection.



> why add the example training data to the repo when you can just get it via `ACEpotentials.example_dataset`?

It is the first 3 structures from the ACEpotentials training data. Its function is to implement tests for correctness of assembly. I just thought that it is so little data that it was better to just add here is as a copy rather than load it over network every time you do tests.