@wcwitt 2023-09-17
Can someone explain the context here? Is the aim to replace JuLIP?

> The new way is way more faster than the old one, between several tens to several hundred

I am fully prepared to believe this is possible, but we should agree on some larger benchmarks. This could become one: https://github.com/ACEsuit/ACEfit.jl/pull/54#issuecomment-1644079554.

The assembly, particularly in parallel, has felt confusingly sluggish to me for some time. I've been attributing it to something in ACE1/JuLIP (perhaps involving GC), where I've been reluctant to dig deeply, but not sure.