@cortner 2023-09-17
Yes we want to kill JuLIP by moving small pieces into ACE packages or into community packages.