@cortner 2023-09-17
I also had the impression that Teemu s weights are I compatible with ours so this may need to be fixed. But that’s a minor issue.