@cortner 2023-09-18
One small remark, to make the tests entirely fair one should remove the cached neighbourlist after each assembly, 
```
rm_nlist(at::Atoms) = delete!(at.data, "nlist:default")
rm_nlist.(data_j)
```
but this doesn't actually seem to make a measurable difference.