@cortner 2023-09-18
The numbers change a bit when you make the model bigger
```
# length(basis) = 4676,
# length(data_j) = 329
 14.784368 seconds (14.64 M allocations: 8.669 GiB, 64.56% gc time)
  3.377460 seconds (18.59 M allocations: 30.495 GiB, 25.45% gc time)

# length(basis) = 12888
# length(data_j) = 329
 26.681512 seconds (25.48 M allocations: 23.483 GiB, 38.05% gc time)
  9.563755 seconds (37.73 M allocations: 83.549 GiB, 23.07% gc time)
```