@wcwitt 2023-09-18
Hm, am I correct that the new assembly routine _does not_ force garbage collection after each structure, unlike the old one? That is likely a factor - perhaps the main factor - particularly for small tests.

We know forcing GC hurts performance, but it has been the only way to avoid crashes (for both distributed and threaded) during larger assemblies.