@cortner 2023-09-18
What if we add a counter and GC every once in a while?