@wcwitt 2023-09-18
Yeah, perhaps. Or I will merge in one of the batching ideas and we can GC after batches.

Still holding out hope this PR fixes the underlying issue somehow - let's see.