@wcwitt 2023-09-18
Just tried the small benchmark again (from the top of this thread):
```
# old (looks insane, but I triple checked)
20.635631 seconds (5.49 M allocations: 786.066 MiB, 97.01% gc time)
# old, without forcing GC
0.605159 seconds (5.49 M allocations: 786.014 MiB, 17.42% gc time)
# new
0.747427 seconds (5.61 M allocations: 1.094 GiB, 17.34% gc time)
```
So for the tiny model the forced GC explains everything.

Same trend, but less definitive for the bigger model:
```
length(basis) = 12888
length(train_old) = 66

# old
82.048485 seconds (31.27 M allocations: 41.684 GiB, 29.60% gc time)
# old, without forcing GC
62.509240 seconds (31.27 M allocations: 41.684 GiB, 3.69% gc time)
# new
49.417405 seconds (45.56 M allocations: 81.406 GiB, 2.45% gc time)
```